TicTacToeServer/src/test/ResponseTest.java
package test;
import socket.Response;
public class ResponseTest {
    /**
     *
     * @param args
     */
    public static void main(String[] args) {
        testDefaultConstructor();
        testParameterizedConstructor();
        testGettersAndSetters();
    }

    /**
     *
     */
    public static void testDefaultConstructor() {
        Response response = new Response();
        // Write assertions to test default constructor
    }


    /**
     *
     */
    public static void testParameterizedConstructor() {
        Response response = new Response(Response.ResponseStatus.FAILURE, "Error message");
        // Write assertions to test parameterized constructor
    }

    /**
     *
     */
    public static void testGettersAndSetters() {
        Response response = new Response();
        // Test all getters and setters
        response.setStatus(Response.ResponseStatus.SUCCESS);
        response.setMessage("Success message");

        // Write assertions to check if getters and setters work as expected
    }
}
